+++ Casks/tiktok-douyin-dl.rb
@@ -1,5 +1,5 @@
 cask "tiktok-douyin-dl" do
-  version "1.8.1"
+  version "1.8.2"
   sha256 "d554658920895c40cef0958ff1f813fd702275efb68bd94811f15b13e53c5266"
 
   url "https://github.com/Francis-Xavier-code/tiktok-douyin-dl/releases/download/v#{version}/MediaDownloader-macOS-#{version}-unsigned.dmg"